pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sensei"
version = "0.0.0"
description = "A pip-installable CLI that scaffolds a learning-environment folder for any LLM agent to operate as an adaptive mentor."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
sensei = "sensei.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sensei"]

[tool.hatch.build.targets.wheel.force-include]
"src/sensei/engine" = "sensei/engine"

[tool.hatch.build.targets.sdist]
include = [
    "src/sensei",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
